Закончил переход на трехслойную архитектуру. В планах реализация сессии и connection pul'а
