--- routes/auth.py
@@ -1,2 +1,0 @@
-# Presentation: роуты
-
